Clean up of rcs_functions module, started to break out the io into their own module
